Refactor main.py for improved readability and efficiency
This commit modifies the main.py for better code readability and efficiency. It changes the structure of the dataframe by extracting column selection and renaming into separate constants and function. This not only simplifies the export() method but also makes the code more maintainable.

Also, a check has been added for the existence of the input file before processing it to avoid crash in case of an absent file.

The creation date is now included in the output file, text wrap property has been added for styling and minor changes are done to better align the cells.

These changes improve the overall structuring and performance of the script, leading to a cleaner and safer code.