Add setup.py file
This commit introduces a new setup.py file for the 'Texniko_Programma' package. The setup.py file is crucial for defining package metadata and facilitates the package's distribution and installation. The metadata includes the package's name, version, author information, and a brief description of the package. Adding a setup.py file is a standard, good practice in Python projects, contributing to better maintainability and ease of installation for users.