Remove unnecessary import and correct typos in comments
The 'copy' module was removed from imports as it was not used in the script, improving the efficiency of the script. Some spelling mistakes in the comments were corrected for better readability. Added code to read 'body.html' for emails to ensure consistent email formatting. Previously the body of the emails was hardcoded into the script, new implementation allows easier changes for the email content. Additional unnecessary whitespaces and lines were removed making the script cleaner.