Moved README content, updated prerequisites and added new email packages
The README.md file from the META directory was deleted and a new README.md was created in the base directory. Changes include an updated and elaborated Python packages list necessary for the script, specifically the "smtplib", "email" and time packages. This rearrangement and added content provide clearer setup instructions.